requirements.txt
apache-airflow~=3.0.2
pykrx
pandas
scikit-learn
boto3
xgboost
joblib
python-dotenv
openai
pymongo
apache-airflow-providers-mongo